refactor: BestAdapter 패키지 이동
