+++ app/src/main/java/com/woowa/banchan/ui/home/best/BestFragment.kt
@@ -13,6 +13,7 @@ import com.woowa.banchan.R
 import com.woowa.banchan.data.remote.dto.BestFood
 import com.woowa.banchan.databinding.FragmentBestBinding
 import com.woowa.banchan.ui.common.uistate.UiState
+import com.woowa.banchan.ui.home.best.adapter.BestAdapter
 import com.woowa.banchan.utils.showToast
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.launchIn
