+++ app/src/main/java/com/woowa/banchan/ui/home/best/adapter/BestAdapter.kt
@@ -1,11 +1,10 @@
-package com.woowa.banchan.ui.home.best
+package com.woowa.banchan.ui.home.best.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.woowa.banchan.R
 import com.woowa.banchan.data.remote.dto.BestFoodCategory
 import com.woowa.banchan.data.remote.dto.FoodItem
 import com.woowa.banchan.databinding.ItemBestHeaderBinding
@@ -13,6 +12,10 @@ import com.woowa.banchan.databinding.ItemBestRecyclerviewBinding
 import com.woowa.banchan.databinding.ItemHomeHeaderBinding
 import com.woowa.banchan.ui.home.adapter.HomeItemAdapter
 import com.woowa.banchan.ui.home.adapter.viewholder.HomeHeaderViewHolder
+import com.woowa.banchan.ui.home.adapter.viewholder.HomeRecyclerViewViewHolder
+import com.woowa.banchan.utils.SUB_HEADER
+import com.woowa.banchan.utils.HOME_HEADER
+import com.woowa.banchan.utils.HOME_ITEM
 
 class BestAdapter : ListAdapter<BestFoodCategory, RecyclerView.ViewHolder>(diffUtil) {
 
@@ -25,14 +28,14 @@ class BestAdapter : ListAdapter<BestFoodCategory, RecyclerView.ViewHolder>(diffU
                     ), parent, false
                 )
             )
-            BEST_HEADER -> BestHeaderViewHolder(
+            SUB_HEADER -> BestHeaderViewHolder(
                 ItemBestHeaderBinding.inflate(
                     LayoutInflater.from(
                         parent.context
                     ), parent, false
                 )
             )
-            else -> BestRecyclerViewViewHolder(
+            else -> HomeRecyclerViewViewHolder(
                 ItemBestRecyclerviewBinding.inflate(
                     LayoutInflater.from(
                         parent.context
@@ -45,8 +48,8 @@ class BestAdapter : ListAdapter<BestFoodCategory, RecyclerView.ViewHolder>(diffU
     override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         when (holder.itemViewType) {
             HOME_HEADER -> (holder as HomeHeaderViewHolder).bind("한 번 주문하면\n두 번 반하는 반찬들", true)
-            BEST_HEADER -> (holder as BestHeaderViewHolder).bind(getItem(position))
-            else -> (holder as BestRecyclerViewViewHolder).bind(getItem(position).items)
+            SUB_HEADER -> (holder as BestHeaderViewHolder).bind(getItem(position))
+            else -> (holder as HomeRecyclerViewViewHolder).bind(getItem(position).items)
         }
     }
 
@@ -55,9 +58,9 @@ class BestAdapter : ListAdapter<BestFoodCategory, RecyclerView.ViewHolder>(diffU
             HOME_HEADER
         else {
             if (position % 2 == 1)
-                BEST_HEADER
+                SUB_HEADER
             else
-                BEST
+                HOME_ITEM
         }
     }
 
@@ -72,9 +75,6 @@ class BestAdapter : ListAdapter<BestFoodCategory, RecyclerView.ViewHolder>(diffU
     }
 
     companion object {
-        const val HOME_HEADER = 1
-        const val BEST_HEADER = 2
-        const val BEST = 3
 
         val diffUtil = object : DiffUtil.ItemCallback<BestFoodCategory>() {
             override fun areItemsTheSame(oldItem: BestFoodCategory, newItem: BestFoodCategory): Boolean {
@@ -92,14 +92,5 @@ class BestHeaderViewHolder(private val binding: ItemBestHeaderBinding) :
     RecyclerView.ViewHolder(binding.root) {
     fun bind(foodCategory: BestFoodCategory) {
         binding.beestFoodCategory = foodCategory
-    }
-}
-
-class BestRecyclerViewViewHolder(private val binding: ItemBestRecyclerviewBinding) :
-    RecyclerView.ViewHolder(binding.root) {
-    fun bind(categoryFood: List<FoodItem>) {
-        val adapter = HomeItemAdapter()
-        adapter.submitList(categoryFood)
-        binding.layoutBest.adapter = adapter
     }
 }
