+++ app/src/main/java/com/woowa/banchan/ui/home/main/MainFragment.kt
@@ -4,15 +4,34 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import com.woowa.banchan.R
+import com.woowa.banchan.databinding.FragmentMainBinding
+import com.woowa.banchan.ui.home.main.adapter.MainAdapter
 
 class MainFragment : Fragment() {
+
+    private lateinit var binding: FragmentMainBinding
+
+    private val mainAdapter: MainAdapter by lazy {
+        MainAdapter()
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_main, container, false)
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_main, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        initAdapter()
+    }
+
+    private fun initAdapter() {
+        binding.layoutMain.adapter = mainAdapter
     }
 }
